# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Oficinas Mecânicas management system
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Oficina {
  id: number;
  nome: string;
  endereco: string;
  especialidades: string;
  createdAt: string;
}

export interface OficinaInput {
  /** @minLength 1 */
  nome: string;
  /** @minLength 1 */
  endereco: string;
  /** @minLength 1 */
  especialidades: string;
}

export interface Veiculo {
  id: number;
  placa: string;
  modelo: string;
  ano: string;
  proprietario: string;
  /** @nullable */
  oficinaId?: number | null;
  createdAt: string;
}

export interface OficinaWithVeiculos {
  id: number;
  nome: string;
  endereco: string;
  especialidades: string;
  createdAt: string;
  veiculos: Veiculo[];
}

export interface VeiculoInput {
  /** @minLength 1 */
  placa: string;
  /** @minLength 1 */
  modelo: string;
  /** @minLength 1 */
  ano: string;
  /** @minLength 1 */
  proprietario: string;
  /** @nullable */
  oficinaId?: number | null;
}

export interface Manutencao {
  id: number;
  descricao: string;
  data: string;
  veiculoId: number;
  /** @nullable */
  oficinaNome?: string | null;
  createdAt: string;
}

export interface VeiculoWithManutencoes {
  id: number;
  placa: string;
  modelo: string;
  ano: string;
  proprietario: string;
  /** @nullable */
  oficinaId?: number | null;
  createdAt: string;
  manutencoes: Manutencao[];
}

export interface DashboardStats {
  totalOficinas: number;
  totalVeiculos: number;
  totalManutencoes: number;
}

export interface VeiculosPorAno {
  ano: string;
  total: number;
}

export interface ErrorResponse {
  error: string;
}