# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Oficinas Mecânicas management system
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalOficinas: number;
  totalVeiculos: number;
  totalManutencoes: number;
}