# lib/api-zod/src/generated/types/veiculoInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Oficinas Mecânicas management system
 * OpenAPI spec version: 0.1.0
 */

export interface VeiculoInput {
  /** @minLength 1 */
  placa: string;
  /** @minLength 1 */
  modelo: string;
  /** @minLength 1 */
  ano: string;
  /** @minLength 1 */
  proprietario: string;
  /** @nullable */
  oficinaId?: number | null;
}