# lib/api-zod/src/generated/types/veiculoWithManutencoes.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Oficinas Mecânicas management system
 * OpenAPI spec version: 0.1.0
 */
import type { Manutencao } from './manutencao';

export interface VeiculoWithManutencoes {
  id: number;
  placa: string;
  modelo: string;
  ano: string;
  proprietario: string;
  /** @nullable */
  oficinaId?: number | null;
  createdAt: Date;
  manutencoes: Manutencao[];
}